requirements.txt
Appium-Python-Client
requests
PyYAML